Move parser test to tests/
